refactor: rename the hole project
BREAKING CHANGE
closes #3
